created a questions folder and added a img tto the app.js
