+++ code/src/App.js
@@ -2,14 +2,16 @@ import React from "react";
 
 import Survey from "./Survey";
 
+import CatPaws from "img/cat-paws.png";
 
 
 export const App = () => {
  
   return (
-    <div>
+    <div className="app-container">
       <Survey />
-      </div>
+       <img className="paws" src= {CatPaws} alt="cat with bouncing tail"/>
+    </div>
   )
 }
