+++ code/src/Questions/Select.js
@@ -3,9 +3,9 @@ import React from 'react';
 const Select = ( {fur, setFur} ) => { 
     
     return ( 
-      <section className="fur-wrapper">      
+      <section>      
         <label htmlFor="fur">
-        <select id="fur" name="fur"
+          <select id="fur" name="fur"
             onChange = {event => setFur(event.target.value)}
             value = {fur} 
             required >
@@ -13,10 +13,9 @@ const Select = ( {fur, setFur} ) => {
               <option value = "short fur" > Short fur</option>
               <option value = "long fur" > Long fur</option>
               <option value = "no fur " > No fur </option>
-        </select>
+          </select>
         </label>
-      </section>
-      
+      </section> 
     )
   }
