+++ code/src/Welcome.js
@@ -1,10 +1,15 @@
 import React from 'react'
 
+
+
 export const Welcome = () => { 
     return (
         <>
+            <section className="welcome">
             <h1 tabIndex ="0">Dream Cat</h1>
             <p tabIndex ="0">Take this survey to see what your dream cat is!</p>
+            </section>
+            
         </>
     )
     }
